chore: move resources
