+++ packages/cli/src/resources/common/config.ts
