+++ packages/cli/src/resources/common/debug.ts
