+++ packages/cli/src/resources/resources.ts
