+++ packages/cli/tsconfig.json
@@ -8,7 +8,7 @@
     "composite": true
   },
   "include": ["src/**/*.ts"],
-  "exclude": ["dist", "node_modules", "src/**/*.test.ts"],
+  "exclude": ["dist", "node_modules"],
   "references": [
     { "path": "../atlas-telemetry" },
     { "path": "../atlas-api-client" },
