+++ packages/integration-tests/src/tools/mongodb/read/export.test.ts
@@ -12,7 +12,7 @@ import {
 } from "../../../integrationHelpers.js";
 import { describeWithMongoDB } from "../../../mongodbHelpers.js";
 import type { UserConfig } from "mongodb-mcp-server";
-import { ExportedData } from "mongodb-mcp-server";
+import { ExportedData } from "@mongodb-js/mcp-cli";
 
 const userConfig: UserConfig = {
     ...defaultTestConfig,
